+++ mail/views/message.php
@@ -30,14 +30,22 @@ if ( 'text' == strtolower( $msg->BodyType)) {
         $_msg = mb_convert_encoding( $msg->Body, 'ASCII', 'HTML-ENTITIES');
         // $_msg = $msg->Body;
         // \sys::logger( sprintf('<%s> %s', 'no conversion', __METHOD__));
-        // $f = sprintf('%s/temp-4-message.html', \config::dataPath());
+        // $f = sprintf('%s/temp-4-message.txt', \config::dataPath());
         // if ( \file_exists($f)) unlink( $f);
         // \file_put_contents( $f, $_msg);
 
 
     }
+    elseif ( !$encoding) {
+        // there is no encoding
+        $_msg = $msg->Body;
+
+    }
     else {
-        $_msg = sprintf( "Encoding: %s\n", $encoding, $msg->Body);
+        $_msg = sprintf( "Encoding: %s\n%s", $encoding, $msg->Body);
+        // $f = sprintf('%s/temp-3-message.txt', \config::dataPath());
+        // if ( \file_exists($f)) unlink( $f);
+        // \file_put_contents( $f, $_msg);
 
     }